# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Food Roulette API
  description: 'Contract for the Flutter client. Frozen per ROADMAP §3 — additive
    changes are safe, renames are not.


    Every non-2xx response uses the ErrorEnvelope shape.'
  version: 0.1.0
paths:
...
  /restaurants:
    get:
      summary: List Restaurants
      description: All restaurants with a count of their menu items (smoke-test endpoint).
      operationId: list_restaurants_restaurants_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  type: object
                type: array
                title: Response List Restaurants Restaurants Get
        '422':
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /restaurants/{restaurant_id}:
    get:
      summary: Get Restaurant
      description: A restaurant plus its full menu (used to eyeball the ingested data).
      operationId: get_restaurant_restaurants__restaurant_id__get
      parameters:
      - name: restaurant_id
        in: path
        required: true
        schema:
          type: string
          title: Restaurant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Restaurant Restaurants  Restaurant Id  Get
        '422':
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/session:
    post:
      summary: Create Session
      description: 'Fake login for now. Hands out an ID and a token, but nothing checks
        the

        token yet — real login comes later.'
      operationId: create_session_auth_session_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
        '422':
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users/{user_id}/preferences:
    get:
      summary: Get Preferences
      description: 'Look up a user''s saved allergy/diet settings. If they''ve never
        saved any,

        hand back empty defaults instead of an error.'
      operationId: get_preferences_users__user_id__preferences_get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesOut'
        '422':
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      summary: Put Preferences
      description: Save (or overwrite) a user's allergy/diet settings.
      operationId: put_preferences_users__user_id__preferences_put
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferencesIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesOut'
        '422':
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /recommend:
    post:
      summary: Recommend
      description: 'Pick restaurants for this user. First we throw out every dish
        that

        breaks a hard rule (allergy, halal, etc.) — no exceptions. Then, among

        what''s left, we sort by how well it matches soft preferences like spice

        level and distance.'
      operationId: recommend_recommend_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendOut'
        '422':
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /chat:
    post:
      summary: Chat
      description: 'Read the message, filter, then describe what survived (DESIGN
        §6).


        Gemini sits on both ends and never in the middle: it turns the message into

        preferences, and later writes the reply, but the choice of restaurants is

        made by the same filter /recommend uses. If Gemini is unreachable the

        request still succeeds — the user gets their saved-profile recommendations

        and a `degraded` code explaining why the reply is plain.'
      operationId: chat_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatOut'
        '422':
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          description: Error envelope (ROADMAP §3.2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  schemas:
    ChatIn:
      properties:
        user_id:
          type: string
          title: User Id
        text:
          type: string
          title: Text
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        limit:
          type: integer
          maximum: 50.0
          minimum: 1.0
          title: Limit
          default: 10
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
      type: object
      required:
      - user_id
      - text
      title: ChatIn
    ChatOut:
      properties:
        reply:
          type: string
          title: Reply
        recommendations:
          items:
            $ref: '#/components/schemas/RecommendedRestaurant'
          type: array
          title: Recommendations
        session_id:
          type: string
          title: Session Id
        degraded:
          anyOf:
          - type: string
            enum:
            - NO_RESULTS
            - INVALID_PREFS
            - NOT_FOUND
            - LLM_UNAVAILABLE
            - UPSTREAM_TIMEOUT
            - INTERNAL
          - type: 'null'
          title: Degraded
      type: object
      required:
      - reply
      - recommendations
      - session_id
      title: ChatOut
      description: 'Same cards as /recommend, plus the prose that introduces them.


        The client renders cards from `recommendations` and never by reading

        `reply` — that separation is what stops a chatty model from inventing a

        restaurant (DESIGN §6).'
...
    HardConstraints:
      properties:
        allergens:
          items:
            type: string
            enum:
            - peanuts
            - tree_nuts
            - shellfish
            - fish
            - wheat
            - soy
            - milk
            - eggs
            - sesame
          type: array
          title: Allergens
          default: []
        halal:
          type: boolean
          title: Halal
          default: false
        no_beef:
          type: boolean
          title: No Beef
          default: false
        no_pork:
          type: boolean
          title: No Pork
          default: false
        vegetarian:
          type: boolean
          title: Vegetarian
          default: false
        vegan:
          type: boolean
          title: Vegan
          default: false
        jain:
          type: boolean
          title: Jain
          default: false
        jay:
          type: boolean
          title: Jay
          default: false
        celiac:
          type: boolean
          title: Celiac
          default: false
      type: object
      title: HardConstraints
      description: 'Non-negotiable rules. Every one of these is enforced in SQL by

        `filter.py` — the LLM never gets a say (DESIGN §4).'
    PreferencesIn:
      properties:
        hard:
          $ref: '#/components/schemas/HardConstraints'
          default:
            allergens: []
            halal: false
            no_beef: false
            no_pork: false
            vegetarian: false
            vegan: false
            jain: false
            jay: false
            celiac: false
        soft:
          $ref: '#/components/schemas/SoftPreferences'
          default:
            spicy_tolerance: 3
            diet_style: none
            wants_parking: false
      type: object
      title: PreferencesIn
    PreferencesOut:
      properties:
        hard:
          $ref: '#/components/schemas/HardConstraints'
          default:
            allergens: []
            halal: false
            no_beef: false
            no_pork: false
            vegetarian: false
            vegan: false
            jain: false
            jay: false
            celiac: false
        soft:
          $ref: '#/components/schemas/SoftPreferences'
          default:
            spicy_tolerance: 3
            diet_style: none
            wants_parking: false
        user_id:
          type: string
          title: User Id
      type: object
      required:
      - user_id
      title: PreferencesOut
    RecommendIn:
      properties:
        user_id:
          type: string
          title: User Id
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        limit:
          type: integer
          maximum: 50.0
          minimum: 1.0
          title: Limit
          default: 10
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
      type: object
      required:
      - user_id
      title: RecommendIn
    RecommendOut:
      properties:
        recommendations:
          items:
            $ref: '#/components/schemas/RecommendedRestaurant'
          type: array
          title: Recommendations
        fallback_used:
          type: boolean
          title: Fallback Used
      type: object
      required:
      - recommendations
      - fallback_used
      title: RecommendOut
    RecommendedRestaurant:
      properties:
        restaurant_id:
          type: string
          title: Restaurant Id
        name_th:
          anyOf:
          - type: string
          - type: 'null'
          title: Name Th
        name_en:
          anyOf:
          - type: string
          - type: 'null'
          title: Name En
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        distance_m:
          anyOf:
          - type: number
          - type: 'null'
          title: Distance M
        price_tier:
          anyOf:
          - type: string
          - type: 'null'
          title: Price Tier
        rating:
          anyOf:
          - type: number
          - type: 'null'
          title: Rating
        photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Url
        safety_tier:
          type: string
          enum:
          - verified
          - unverified
          title: Safety Tier
        needs_ack:
          type: boolean
          title: Needs Ack
        ack_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Ack Reason
        excluded_count:
          type: integer
          title: Excluded Count
        safe_dishes:
          items:
            $ref: '#/components/schemas/SafeDish'
          type: array
          title: Safe Dishes
      type: object
      required:
      - restaurant_id
      - name_th
      - name_en
      - latitude
      - longitude
      - distance_m
      - price_tier
      - safety_tier
      - needs_ack
      - excluded_count
      - safe_dishes
      title: RecommendedRestaurant
      description: 'The single object the whole UI is built from (ROADMAP §3.1).


        `rating` and `photo_url` stay null until M4 wires up Google Places; the

        client must render gracefully without them from day one.'
...
    SessionIn:
      properties:
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
      type: object
      title: SessionIn
    SessionOut:
      properties:
        user_id:
          type: string
          title: User Id
        token:
          type: string
          title: Token
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
      type: object
      required:
      - user_id
      - token
      title: SessionOut
    SoftPreferences:
      properties:
        spicy_tolerance:
          type: integer
          maximum: 3.0
          minimum: 0.0
          title: Spicy Tolerance
          default: 3
        price_tier:
          anyOf:
          - type: string
            enum:
            - $
            - $$
            - $$$
            - $$$$
          - type: 'null'
          title: Price Tier
        diet_style:
          type: string
          enum:
          - keto
          - low_carb
          - mediterranean
          - carnivore
          - none
          title: Diet Style
          default: none
        wants_parking:
          type: boolean
          title: Wants Parking
          default: false
      type: object
      title: SoftPreferences
...
    ErrorBody:
      properties:
        code:
          enum:
          - NO_RESULTS
          - INVALID_PREFS
          - NOT_FOUND
          - LLM_UNAVAILABLE
          - UPSTREAM_TIMEOUT
          - INTERNAL
          title: Code
          type: string
        message:
          title: Message
          type: string
        detail:
          anyOf:
          - type: string
          - type: 'null'
          default: null
          title: Detail
      required:
      - code
      - message
      title: ErrorBody
      type: object
    ErrorEnvelope:
      $defs:
        ErrorBody:
          properties:
            code:
              enum:
              - NO_RESULTS
              - INVALID_PREFS
              - NOT_FOUND
              - LLM_UNAVAILABLE
              - UPSTREAM_TIMEOUT
              - INTERNAL
              title: Code
              type: string
            message:
              title: Message
              type: string
            detail:
              anyOf:
              - type: string
              - type: 'null'
              default: null
              title: Detail
          required:
          - code
          - message
          title: ErrorBody
          type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      required:
      - error
      title: ErrorEnvelope
      type: object